avoid encoding/decoding base64 as much as possible
base64 en/decoding is rather slow and there were places where it was being
done rather wastefully. For example, when requesting the key from the server
it would be sent in Base64, and then decoded into binary, only to be encoded
back into base64 to be stored. To avoid this we store the binary where we can
and the only place we truely need to use base64 is when stompd is sending
the data to the client