implement hooks functionality
a hooks file is now installed and can have code added to it that will be
run after each command. The command is passed in so that code can be run
depending on which command is used, eg. a hook can be added that runs
git add+commit+push whenever `stomp add/edit/remove` is run